+++ my-admin/src/components/LoginForm/LoginForm.scss
@@ -1,0 +1,58 @@
+@import "../../variables.scss";
+
+.login-form {
+  width: 384px;
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+  position: relative;
+
+  &::after {
+    content: url("../../assets/svg/icon-eye.svg");
+    display: flex;
+    width: 20px;
+    height: 20px;
+    position: absolute;
+    top: 142px;
+    left: 348px;
+    cursor: pointer;
+  }
+
+  &__label {
+    font-size: 14px;
+    font-weight: 600;
+    line-height: 20px;
+  }
+
+  &__input {
+    width: 100%;
+    height: 48px;
+    margin-top: 8px;
+    border-radius: 8px;
+    border: 1px solid #ebebed;
+    padding: 14px 16px;
+
+    &::placeholder {
+      color: #b1b5c3;
+      font-size: 14px;
+      line-height: 20px;
+    }
+  }
+
+  &__submit {
+    height: 48px;
+    margin-top: 84px;
+    display: flex;
+    padding: 12px 120px;
+    justify-content: center;
+    align-items: center;
+    border-radius: 8px;
+    background-color: $logo-color;
+    border: none;
+    font-size: 14px;
+    font-weight: 600;
+    line-height: 24px;
+    color: white;
+    cursor: pointer;
+  }
+}
