+++ my-admin/src/components/LoginForm/LoginForm.tsx
@@ -1,0 +1,45 @@
+import { useLogin, useNotify } from "react-admin";
+import { useState } from "react";
+import "./LoginForm.scss";
+
+const LoginForm = () => {
+  const [email, setEmail] = useState<string>("");
+  const [password, setPassword] = useState<string>("");
+  const login = useLogin();
+  const notify = useNotify();
+
+  const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+    login({ email, password }).catch(() => notify("Invalid email or password"));
+  };
+
+  return (
+    <form onSubmit={handleSubmit} className="login-form">
+      <label htmlFor="email" className="login-form__label">
+        Email
+        <input
+          className="login-form__input"
+          name="email"
+          type="email"
+          placeholder="Enter your email"
+          value={email}
+          onChange={(e) => setEmail(e.target.value)}
+        />
+      </label>
+      <label htmlFor="password" className="login-form__label">
+        Password
+        <input
+          className="login-form__input"
+          name="password"
+          type="password"
+          placeholder="Enter your password"
+          value={password}
+          onChange={(e) => setPassword(e.target.value)}
+        />
+      </label>
+      <input type="submit" className="login-form__submit" value="Login" />
+    </form>
+  );
+};
+
+export default LoginForm;
