+++ my-admin/src/pages/LoginPage/LoginPage.scss
@@ -3,82 +3,25 @@
 .login {
   display: flex;
 
-  &__form {
+  &__wrap {
     width: 50%;
     height: 100vh;
     padding: 114px 0 0 112px;
+  }
 
-    &__title {
-      font-size: 32px;
-      font-weight: 800;
-      line-height: 40px;
-      margin: 0 0 8px 0;
-    }
+  &__title {
+    font-size: 32px;
+    font-weight: 800;
+    line-height: 40px;
+    margin: 0 0 8px 0;
+  }
 
-    &__text {
-      width: 357px;
-      font-size: 14px;
-      line-height: 20px;
-      font-weight: 600;
-      margin: 0 0 48px 0;
-    }
-
-    &__wrap {
-      width: 384px;
-      display: flex;
-      flex-direction: column;
-      gap: 24px;
-      position: relative;
-
-      &::after {
-        content: url("../../assets/svg/icon-eye.svg");
-        display: flex;
-        width: 20px;
-        height: 20px;
-        position: absolute;
-        top: 142px;
-        left: 348px;
-        cursor: pointer;
-      }
-    }
-
-    &__label {
-      font-size: 14px;
-      font-weight: 600;
-      line-height: 20px;
-    }
-
-    &__input {
-      width: 100%;
-      height: 48px;
-      margin-top: 8px;
-      border-radius: 8px;
-      border: 1px solid #ebebed;
-      padding: 14px 16px;
-
-      &::placeholder {
-        color: var(--gray-backsground-gray, #b1b5c3);
-        font-size: 14px;
-        line-height: 20px;
-      }
-    }
-
-    &__submit {
-      height: 48px;
-      margin-top: 84px;
-      display: flex;
-      padding: 12px 120px;
-      justify-content: center;
-      align-items: center;
-      border-radius: 8px;
-      background-color: $logo-color;
-      border: none;
-      font-size: 14px;
-      font-weight: 600;
-      line-height: 24px;
-      color: white;
-      cursor: pointer;
-    }
+  &__text {
+    width: 357px;
+    font-size: 14px;
+    line-height: 20px;
+    font-weight: 600;
+    margin: 0 0 48px 0;
   }
 
   &__logo {
